Port to .NET Core and make tests pass
