--- KenticoCloudDotNetGenerators.Tests/packages.config
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="KenticoCloud.Delivery" version="3.0.1" targetFramework="net452" />
-  <package id="Microsoft.CodeAnalysis" version="1.3.2" targetFramework="net452" />
-  <package id="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" targetFramework="net452" />
-  <package id="Microsoft.CodeAnalysis.Common" version="1.3.2" targetFramework="net452" />
-  <package id="Microsoft.CodeAnalysis.CSharp" version="1.3.2" targetFramework="net452" />
-  <package id="Microsoft.CodeAnalysis.CSharp.Workspaces" version="1.3.2" targetFramework="net452" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic" version="1.3.2" targetFramework="net452" />
-  <package id="Microsoft.CodeAnalysis.VisualBasic.Workspaces" version="1.3.2" targetFramework="net452" />
-  <package id="Microsoft.CodeAnalysis.Workspaces.Common" version="1.3.2" targetFramework="net452" />
-  <package id="Microsoft.Composition" version="1.0.27" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net452" />
-  <package id="NUnit" version="3.6.0" targetFramework="net452" />
-  <package id="NUnit3TestAdapter" version="3.7.0" targetFramework="net452" />
-  <package id="System.Collections" version="4.0.0" targetFramework="net452" />
-  <package id="System.Collections.Immutable" version="1.1.37" targetFramework="net452" />
-  <package id="System.Diagnostics.Debug" version="4.0.0" targetFramework="net452" />
-  <package id="System.Globalization" version="4.0.0" targetFramework="net452" />
-  <package id="System.Linq" version="4.0.0" targetFramework="net452" />
-  <package id="System.Reflection.Metadata" version="1.2.0" targetFramework="net452" />
-  <package id="System.Resources.ResourceManager" version="4.0.0" targetFramework="net452" />
-  <package id="System.Runtime" version="4.0.0" targetFramework="net452" />
-  <package id="System.Runtime.Extensions" version="4.0.0" targetFramework="net452" />
-  <package id="System.Threading" version="4.0.0" targetFramework="net452" />
-</packages>
