--- KenticoCloudDotNetGenerators.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KenticoCloudDotNetGenerators", "KenticoCloudDotNetGenerators\KenticoCloudDotNetGenerators.csproj", "{ACC333BF-1542-4C1B-9442-45DC741B843F}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KenticoCloudDotNetGenerators.Tests", "KenticoCloudDotNetGenerators.Tests\KenticoCloudDotNetGenerators.Tests.csproj", "{308DCEAF-257C-4A1B-85F6-5610EF26072B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{ACC333BF-1542-4C1B-9442-45DC741B843F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ACC333BF-1542-4C1B-9442-45DC741B843F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ACC333BF-1542-4C1B-9442-45DC741B843F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ACC333BF-1542-4C1B-9442-45DC741B843F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{308DCEAF-257C-4A1B-85F6-5610EF26072B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{308DCEAF-257C-4A1B-85F6-5610EF26072B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{308DCEAF-257C-4A1B-85F6-5610EF26072B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{308DCEAF-257C-4A1B-85F6-5610EF26072B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
