--- KenticoCloudDotNetGenerators/Program.cs
@@ -1,14 +1,0 @@
-﻿namespace KenticoCloudDotNetGenerators
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            var options = new Options();
-            if (CommandLine.Parser.Default.ParseArguments(args, options))
-            {
-                new CodeGenerator(options).Run();
-            }
-        }
-    }
-}
