+++ project.json
@@ -1,0 +1,25 @@
+{
+  "version": "1.0.0-*",
+  "buildOptions": {
+    "debugType": "portable",
+    "emitEntryPoint": true
+  },
+  "dependencies": {
+    "KenticoCloud.Delivery": "3.0.2",
+    "Microsoft.CodeAnalysis": "1.3.2",
+    "NUnit": "3.6.1",
+    "dotnet-test-nunit": "3.4.0-beta-3"
+  },
+  "testRunner": "nunit",
+  "frameworks": {
+    "netcoreapp1.1": {
+      "dependencies": {
+        "Microsoft.NETCore.App": {
+          "type": "platform",
+          "version": "1.1.0"
+        }
+      },
+      "imports": "portable-net45+win8+wp8+wpa81"
+    }
+  }
+}
