+++ src/ClassCodeGenerator.cs
