+++ src/ClassDefinition.cs
