+++ src/InvalidIdentifierException.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace KenticoCloudDotNetGenerators
+{
+    public class InvalidIdentifierException : Exception
+    {
+        public InvalidIdentifierException()
+        {
+        }
+
+        public InvalidIdentifierException(string message) : base(message)
+        {
+        }
+    }
+}
