+++ src/TextHelpers.cs
