+++ test/ClassDefinitionTests.cs
