+++ test/TextHelpersTests.cs
